fix compilation

+++ assembly/contracts/autonomous.ts
@@ -49,7 +49,7 @@ function generateRandomIncrease(base: u64): u64 {
   const increasePercent = abs(randomInt) % 10 - 5;
   const increase = base as i64 * increasePercent / 100;
 
-  if(increase < 0 && base <= abs(increase)) {
+  if(increase < 0 && base <= (abs(increase) as u64)) {
     return 0;
   }
   return base + increase;
